6.ADDED lib folder with getFunctions, init, prompts, and viewFunctions
