+++ lib/init.js
@@ -1,25 +1,7 @@
-const mysql = require ("mysql2");
-const inquirer = require ("inquirer");
-const { prompts } = require ("./lib/prompts.js");
-const { viewDepartments, viewRoles, viewEmployees } = require ("./lib/viewFunctions.js")
-
-const db = mysql.createConnection(
-    {
-        host:"localhost",
-        user:"root",
-        password:"password",
-        database: "company_db"
-    },
-)
-
-db.connect((error) => {
-    if (error) {
-      console.error("Error connecting to the database:", error);
-      return;
-    }
-    console.log("Connected to the company_db database.");
-    init();
-});
+const inquirer = require("inquirer");
+const { prompts } = require("./prompts.js");
+const { getDepartments, getRoles, getEmployees, getManagers} = require ("./getFunctions.js");
+const { viewDepartments, viewRoles, viewEmployees } = require("./viewFunctions.js");
 
 function init() {
     inquirer.prompt(prompts).then((answers) => {
@@ -44,6 +26,8 @@ function init() {
       // }     
     });
   }
+
+  module.exports = {init}
 
     // TODO database manipulation
         // Add
